tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-HTMLObjectElement.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="require-trusted-types">
</head>
<body>
<script>
  //helper function for the tests
  function testData(str, url) {
    var objectElement = document.createElement('object');
    objectElement.data = url;
    objectElement.codeBase = url;
    assert_equals(objectElement.data, str);
    assert_equals(objectElement.codeBase, str);
  }

  //URL assignments do not throw
  test(t => {
    testData(URLS.safe, TrustedURL.create(URLS.safe));
  }, "Basic processing: safe URL, safe construction.");

  test(t => {
    testData(URLS.safe, TrustedURL.unsafelyCreate(URLS.safe));
  }, "Basic processing: safe URL, unsafe construction.");

  //String assignments throw
  test(t => {
    var objectElement = document.createElement('object');
    assert_throws(new TypeError(), _ => {
      objectElement.data = "A string";
    });
  }, "`objectElement.data = string` throws");

  test(t => {
    var objectElement = document.createElement('object');
    assert_throws(new TypeError(), _ => {
      objectElement.codeBase = "A string";
    });
  }, "`objectElement.codeBase = string` throws");

  //Null assignment throws.
  test(t => {
    var objectElement = document.createElement('object');
    assert_throws(new TypeError(), _ => {
      objectElement.data = null;
    });
  }, "`objectElement.data = null` throws");

  //Null assignment throws.
  test(t => {
    var objectElement = document.createElement('object');
    assert_throws(new TypeError(), _ => {
      objectElement.codeBase = null;
    });
  }, "`objectElement.codeBase = null` throws");
</script>
